requirements.txt
django~=2.2
django-cleanup~=3.2
django-cors-headers~=2.5
django-enumfield~=1.5
django-storages~=1.7
djangorestframework~=3.9
djangorestframework-jwt~=1.11
dj-database-url~=0.5
pillow~=6.0
python-decouple~=3.1
sorl-thumbnail~=12.5